+++ packages/nimble-components/src/rich-text/editor/tests/rich-text-editor-mention.spec.ts
@@ -1173,8 +1173,7 @@ describe('RichTextEditorMentionListbox', () => {
             expect(pageObject.getSelectedOption()).toBe('user2 name');
         });
 
-        // Intermittent, see https://github.com/ni/nimble/issues/2150
-        xit('should commit mention into the editor when clicked', async () => {
+        it('should commit mention into the editor when clicked', async () => {
             await appendUserMentionConfiguration(element, [
                 { key: 'user:1', displayName: 'username1' }
             ]);
@@ -1193,8 +1192,7 @@ describe('RichTextEditorMentionListbox', () => {
             expect(pageObject.isMentionListboxOpened()).toBeFalse();
         });
 
-        // Intermittent, see https://github.com/ni/nimble/issues/2150
-        xit('should commit mention into the editor on Enter', async () => {
+        it('should commit mention into the editor on Enter', async () => {
             await appendUserMentionConfiguration(element, [
                 { key: 'user:1', displayName: 'username1' }
             ]);
@@ -1213,8 +1211,7 @@ describe('RichTextEditorMentionListbox', () => {
             expect(pageObject.isMentionListboxOpened()).toBeFalse();
         });
 
-        // Intermittent, see https://github.com/ni/nimble/issues/2150
-        xit('should commit mention into the editor on Tab', async () => {
+        it('should commit mention into the editor on Tab', async () => {
             await appendUserMentionConfiguration(element, [
                 { key: 'user:1', displayName: 'username1' }
             ]);
@@ -1245,8 +1242,7 @@ describe('RichTextEditorMentionListbox', () => {
             expect(pageObject.isMentionListboxOpened()).toBeFalse();
         });
 
-        // Intermittent, see https://github.com/ni/nimble/issues/2150
-        xit('should filter and commit first mention into the editor on Enter', async () => {
+        it('should filter and commit first mention into the editor on Enter', async () => {
             await appendUserMentionConfiguration(element, [
                 { key: 'user:1', displayName: 'username1' },
                 { key: 'user:2', displayName: 'username2' }
@@ -1266,8 +1262,7 @@ describe('RichTextEditorMentionListbox', () => {
             ).toEqual(['username2']);
         });
 
-        // Intermittent, see https://github.com/ni/nimble/issues/2150
-        xit('should filter and commit first mention into the editor on Tab', async () => {
+        it('should filter and commit first mention into the editor on Tab', async () => {
             await appendUserMentionConfiguration(element, [
                 { key: 'user:1', displayName: 'username1' },
                 { key: 'user:2', displayName: 'username2' }
@@ -1547,8 +1542,7 @@ describe('RichTextEditorMentionListbox', () => {
             expect(pageObject.isMentionListboxOpened()).toBeTrue();
         });
 
-        // Intermittent, see https://github.com/ni/nimble/issues/2150
-        xit('should show mention popup for multiple mention configuration elements', async () => {
+        it('should show mention popup for multiple mention configuration elements', async () => {
             await appendUserMentionConfiguration(element, [
                 { key: 'user:1', displayName: 'username1' },
                 { key: 'user:2', displayName: 'username2' }