+++ packages/spright-components/CHANGELOG.md
@@ -1,8 +1,48 @@
 # Change Log - @ni/spright-components
 
-This log was last generated on Mon, 17 Jun 2024 16:38:20 GMT and should not be manually modified.
+This log was last generated on Tue, 25 Jun 2024 18:38:46 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.0.27
+
+Tue, 25 Jun 2024 18:38:46 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v29.7.0
+
+## 1.0.26
+
+Thu, 20 Jun 2024 20:08:12 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v29.6.0
+
+## 1.0.25
+
+Tue, 18 Jun 2024 22:10:11 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v29.5.0
+
+## 1.0.24
+
+Tue, 18 Jun 2024 18:31:52 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v29.4.2
+
+## 1.0.23
+
+Mon, 17 Jun 2024 22:45:32 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v29.4.1
 
 ## 1.0.22